+++ lib/index.ts
@@ -1,5 +1,6 @@
 import './index.css';
 
 export { Arrow } from './react/Arrow';
-export { LineContextProvider, useLineContext } from './react/LineContext';
 export { useReducedGraphics } from './react/useReducedGraphics';
+export { ArrowsContextProvider, useArrowsContext } from './react/ArrowsContext';
+export { ArrowsContainer } from './react/ArrowsContainer';