+++ sandbox/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { App } from './App2.jsx';
+import { App } from './App3.jsx';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('app')!).render(<App />);